requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.3
djangorestframework==3.12.1
idna==2.10
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
Pillow==7.2.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
django-extensions==3.0.9
requests==2.24.0
google-api-python-client==1.12.5
google-auth-oauthlib==0.4.2
django-cors-headers==3.5.0
psycopg2-binary==2.8.6
